MAINT: special: Fix a few warnings that occur when compiling the cephes code.
* Removed the unused variable `err` in the function `hyp2f1_neg_c_equal_bc`
  in hyp2f1.c to fix:

      scipy/special/cephes/hyp2f1.c:607:12: warning: unused variable 'err' [-Wunused-variable]
          double err;
                 ^

* Removed the unused function prototype for `igam_leading_factor` in igam.c
  to fix:

      scipy/special/cephes/igam.c:123:15: warning: unused function 'igam_leading_factor' [-Wunused-function]
      static double igam_leading_factor(double, double);
                    ^

* Add a cast of `code` to `unsigned long` in the function `mtherr` in mtherr.c
  to fix:

      scipy/special/cephes/mtherr.c:83:27: warning: comparison of integers of different signs: 'int' and 'unsigned long' [-Wsign-compare]
          if (code <= 0 || code >= sizeof(conv_to_sf) / sizeof(conv_to_sf[0])) {
                           ~~~~ ^  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

* Add `const` qualifiers where needed in the functions in polevl.h to fix
  several warnings of the form:

      scipy/special/cephes/expn.c:216:28: warning: passing 'const double *' to parameter of type 'double *' discards qualifiers [-Wincompatible-pointer-types-discards-qualifiers]
              term = fac*polevl(lambda, A[k], Adegs[k]);
                                        ^~~~
      scipy/special/cephes/polevl.h:67:50: note: passing argument to parameter 'coef' here
      static NPY_INLINE double polevl(double x, double coef[], int N)

+++ scipy/special/cephes/mtherr.c
@@ -80,7 +80,7 @@ void mtherr(const char *name, int code)
     /* Display error message defined
      * by the code argument.
      */
-    if (code <= 0 || code >= sizeof(conv_to_sf) / sizeof(conv_to_sf[0])) {
+    if (code <= 0 || (unsigned long) code >= sizeof(conv_to_sf) / sizeof(conv_to_sf[0])) {
         code = 0;
     }
